+++ src/RcppExports.cpp
@@ -2292,8 +2292,8 @@ BEGIN_RCPP
 END_RCPP
 }
 // cpp_simulate_distance
-Rcpp::List cpp_simulate_distance(Rcpp::NumericMatrix X_lambda, Rcpp::NumericMatrix X_sigma, Rcpp::NumericMatrix draws, Rcpp::NumericVector cutpoints, int key, int transect, double b_shape, int n_sites, int n_bins, int p_lam, int p_sig, bool is_nb, double r_size, int nsim);
-RcppExport SEXP _tulpaObs_cpp_simulate_distance(SEXP X_lambdaSEXP, SEXP X_sigmaSEXP, SEXP drawsSEXP, SEXP cutpointsSEXP, SEXP keySEXP, SEXP transectSEXP, SEXP b_shapeSEXP, SEXP n_sitesSEXP, SEXP n_binsSEXP, SEXP p_lamSEXP, SEXP p_sigSEXP, SEXP is_nbSEXP, SEXP r_sizeSEXP, SEXP nsimSEXP) {
+Rcpp::List cpp_simulate_distance(Rcpp::NumericMatrix X_lambda, Rcpp::NumericMatrix X_sigma, Rcpp::NumericMatrix draws, Rcpp::NumericVector cutpoints, int key, int transect, int quad_order, double b_shape, int n_sites, int n_bins, int p_lam, int p_sig, bool is_nb, double r_size, int nsim);
+RcppExport SEXP _tulpaObs_cpp_simulate_distance(SEXP X_lambdaSEXP, SEXP X_sigmaSEXP, SEXP drawsSEXP, SEXP cutpointsSEXP, SEXP keySEXP, SEXP transectSEXP, SEXP quad_orderSEXP, SEXP b_shapeSEXP, SEXP n_sitesSEXP, SEXP n_binsSEXP, SEXP p_lamSEXP, SEXP p_sigSEXP, SEXP is_nbSEXP, SEXP r_sizeSEXP, SEXP nsimSEXP) {
 BEGIN_RCPP
     Rcpp::RObject rcpp_result_gen;
     Rcpp::RNGScope rcpp_rngScope_gen;
@@ -2303,6 +2303,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< Rcpp::NumericVector >::type cutpoints(cutpointsSEXP);
     Rcpp::traits::input_parameter< int >::type key(keySEXP);
     Rcpp::traits::input_parameter< int >::type transect(transectSEXP);
+    Rcpp::traits::input_parameter< int >::type quad_order(quad_orderSEXP);
     Rcpp::traits::input_parameter< double >::type b_shape(b_shapeSEXP);
     Rcpp::traits::input_parameter< int >::type n_sites(n_sitesSEXP);
     Rcpp::traits::input_parameter< int >::type n_bins(n_binsSEXP);
@@ -2311,7 +2312,7 @@ BEGIN_RCPP
     Rcpp::traits::input_parameter< bool >::type is_nb(is_nbSEXP);
     Rcpp::traits::input_parameter< double >::type r_size(r_sizeSEXP);
     Rcpp::traits::input_parameter< int >::type nsim(nsimSEXP);
-    rcpp_result_gen = Rcpp::wrap(cpp_simulate_distance(X_lambda, X_sigma, draws, cutpoints, key, transect, b_shape, n_sites, n_bins, p_lam, p_sig, is_nb, r_size, nsim));
+    rcpp_result_gen = Rcpp::wrap(cpp_simulate_distance(X_lambda, X_sigma, draws, cutpoints, key, transect, quad_order, b_shape, n_sites, n_bins, p_lam, p_sig, is_nb, r_size, nsim));
     return rcpp_result_gen;
 END_RCPP
 }
@@ -2575,7 +2576,7 @@ static const R_CallMethodDef CallEntries[] = {
     {"_tulpaObs_cpp_simulate_removal", (DL_FUNC) &_tulpaObs_cpp_simulate_removal, 12},
     {"_tulpaObs_cpp_simulate_dyn_abun", (DL_FUNC) &_tulpaObs_cpp_simulate_dyn_abun, 15},
     {"_tulpaObs_cpp_simulate_fp_occu", (DL_FUNC) &_tulpaObs_cpp_simulate_fp_occu, 12},
-    {"_tulpaObs_cpp_simulate_distance", (DL_FUNC) &_tulpaObs_cpp_simulate_distance, 14},
+    {"_tulpaObs_cpp_simulate_distance", (DL_FUNC) &_tulpaObs_cpp_simulate_distance, 15},
     {"_tulpaObs_cpp_simulate_ms_nmix", (DL_FUNC) &_tulpaObs_cpp_simulate_ms_nmix, 8},
     {"_tulpaObs_cpp_simulate_ms_occu", (DL_FUNC) &_tulpaObs_cpp_simulate_ms_occu, 7},
     {"_tulpaObs_cpp_simulate_ms_occu_cover", (DL_FUNC) &_tulpaObs_cpp_simulate_ms_occu_cover, 10},